[Feat] [Done] [Bishwajeet] Add Swagger documentation for challenge API endpoints
