--- langsuit/app/api/challenges/[challengesId]/route.ts
@@ -1,54 +1,0 @@
-import { eq } from "drizzle-orm";
-import { NextResponse, type NextRequest } from "next/server";
-
-import db from "@/db/drizzle";
-import { challenges } from "@/db/schema";
-import { getIsAdmin } from "@/lib/admin";
-
-export const GET = async (
-  _req: NextRequest,
-  { params }: { params: { challengeId: number } }
-) => {
-  const isAdmin = getIsAdmin();
-  if (!isAdmin) return new NextResponse("Unauthorized.", { status: 401 });
-
-  const data = await db.query.challenges.findFirst({
-    where: eq(challenges.id, params.challengeId),
-  });
-
-  return NextResponse.json(data);
-};
-
-export const PUT = async (
-  req: NextRequest,
-  { params }: { params: { challengeId: number } }
-) => {
-  const isAdmin = getIsAdmin();
-  if (!isAdmin) return new NextResponse("Unauthorized.", { status: 401 });
-
-  const body = (await req.json()) as typeof challenges.$inferSelect;
-  const data = await db
-    .update(challenges)
-    .set({
-      ...body,
-    })
-    .where(eq(challenges.id, params.challengeId))
-    .returning();
-
-  return NextResponse.json(data[0]);
-};
-
-export const DELETE = async (
-  _req: NextRequest,
-  { params }: { params: { challengeId: number } }
-) => {
-  const isAdmin = getIsAdmin();
-  if (!isAdmin) return new NextResponse("Unauthorized.", { status: 401 });
-
-  const data = await db
-    .delete(challenges)
-    .where(eq(challenges.id, params.challengeId))
-    .returning();
-
-  return NextResponse.json(data[0]);
-};
